refactor: 패키지 변경에 대한 의존성 수정
